+++ src/core/generateZodSchema.ts
@@ -117,6 +117,20 @@ export function generateZodSchemaVariableStatement({
       skipParseJSDoc,
       getNamespaceSchemaName,
     });
+
+    if (!skipParseJSDoc) {
+      const jsDocTags = getJSDocTags(node, sourceFile);
+      if (jsDocTags.strict) {
+        schema = f.createCallExpression(
+          f.createPropertyAccessExpression(
+            schema,
+            f.createIdentifier("strict")
+          ),
+          undefined,
+          undefined
+        );
+      }
+    }
   }
 
   if (ts.isTypeAliasDeclaration(node)) {
@@ -368,20 +382,44 @@ function buildZodPrimitive({
 
     // Deal with `Record<>` syntax
     if (identifierName === "Record" && typeNode.typeArguments) {
-      if (
-        typeNode.typeArguments.length !== 2 ||
-        typeNode.typeArguments[0].kind !== ts.SyntaxKind.StringKeyword
-      ) {
-        throw new Error(
-          `Record<${typeNode.typeArguments[0].getText(
-            sourceFile
-          )}, …> are not supported (https://github.com/colinhacks/zod/tree/v3#records)`
+      if (typeNode.typeArguments[0].kind === ts.SyntaxKind.StringKeyword) {
+        // Short version (`z.record(zodType)`)
+        return buildZodSchema(
+          z,
+          "record",
+          [
+            buildZodPrimitive({
+              z,
+              typeNode: typeNode.typeArguments[1],
+              isOptional: false,
+              jsDocTags,
+              sourceFile,
+              isPartial: false,
+              dependencies,
+              getDependencyName,
+              skipParseJSDoc,
+            }),
+          ],
+          zodProperties
         );
       }
+
+      // Expanded version (`z.record(zodType, zodType)`)
       return buildZodSchema(
         z,
         "record",
         [
+          buildZodPrimitive({
+            z,
+            typeNode: typeNode.typeArguments[0],
+            isOptional: false,
+            jsDocTags,
+            sourceFile,
+            isPartial: false,
+            dependencies,
+            getDependencyName,
+            skipParseJSDoc,
+          }),
           buildZodPrimitive({
             z,
             typeNode: typeNode.typeArguments[1],
@@ -613,6 +651,25 @@ function buildZodPrimitive({
         zodProperties
       );
     }
+    if (ts.isPrefixUnaryExpression(typeNode.literal)) {
+      if (
+        typeNode.literal.operator === ts.SyntaxKind.MinusToken &&
+        ts.isNumericLiteral(typeNode.literal.operand)
+      ) {
+        return buildZodSchema(
+          z,
+          "literal",
+          [
+            f.createPrefixUnaryExpression(
+              ts.SyntaxKind.MinusToken,
+              f.createNumericLiteral(typeNode.literal.operand.text)
+            ),
+          ],
+          zodProperties
+        );
+      }
+    }
+
     if (typeNode.literal.kind === ts.SyntaxKind.TrueKeyword) {
       return buildZodSchema(z, "literal", [f.createTrue()], zodProperties);
     }